chore: code cleanup in albums page header
