+++ MusicApp/Models/AlbumsPageHeaderViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  AlbumsPageHeaderViewModel.swift
+//  MusicApp
+//
+//  Created by Damien L Thompson on 2024-10-24.
+//
+
+import Foundation
+
+struct AlbumsPageHeaderViewModel {
+
+    let title: String
+    let artisName: String
+    let coverImage: URL?
+    let genre: [String]?
+    let type: ItemType
+}
